requirements.txt
flask==2.2.5
tqdm==4.65.0
Pillow==10.1.0
transformers==4.41.2
streamlit==1.35.0
streamlit-feedback==0.1.3
torch==2.1.2
torchvision==0.16.2
sentencepiece==0.1.99